+++ src/main/java/com/example/demo/repository/ConsultantRepository.java
@@ -4,6 +4,6 @@ package com.example.demo.repository;
 import com.example.demo.entities.Consultant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultantRepository extends JpaRepository<Consultant,Integer> {
+public interface ConsultantRepository extends JpaRepository<Consultant,Long> {
 
 }
