+++ src/main/java/com/example/demo/repository/EntrepriseRepository.java
@@ -5,7 +5,7 @@ import com.example.demo.entities.Entreprise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface EntrepriseRepository extends JpaRepository<Entreprise,Integer> {
+public interface EntrepriseRepository extends JpaRepository<Entreprise,Long> {
 
 }
